Move tests from ./test to alongside files in form *.spec.js
