+++ app/step-mods/PgtSet.spec.js
@@ -5,7 +5,7 @@
 
 const docx = require('docx');
 
-const StepModuleTester = require('./helpers/StepModuleTester');
+const StepModuleTester = require('../../test/helpers/StepModuleTester');
 
 const goodSettings = [
 	{
