+++ app/writer/procedure/ProcedureWriter.spec.js
@@ -7,14 +7,14 @@ const childProcess = require('child_process');
 const path = require('path');
 const assert = require('chai').assert;
 
-const CommanderProgram = require('../app/model/CommanderProgram');
-const Procedure = require('../app/model/Procedure');
+const CommanderProgram = require('../../model/CommanderProgram');
+const Procedure = require('../../model/Procedure');
 
-const EvaDocxProcedureWriter = require('../app/writer/procedure/EvaDocxProcedureWriter');
+const EvaDocxProcedureWriter = require('./EvaDocxProcedureWriter');
 
 describe('ProcedureWriter', function() {
 	const procedure = new Procedure();
-	const procedureFile = path.join(__dirname, 'cases/simple/procedures/proc.yml');
+	const procedureFile = path.join(__dirname, '../../../test/cases/simple/procedures/proc.yml');
 
 	const err = procedure.addProcedureDefinitionFromFile(procedureFile);
 	if (err) {
